Move SQL code from client to server
