+++ sql/driver/conn.go
@@ -18,24 +18,16 @@
 package driver
 
 import (
-	"bytes"
 	"database/sql/driver"
 	"errors"
 	"fmt"
 	"math"
-	"strings"
-	"time"
 
 	"golang.org/x/net/context"
 
 	"github.com/cockroachdb/cockroach/client"
-	"github.com/cockroachdb/cockroach/keys"
-	"github.com/cockroachdb/cockroach/proto"
 	"github.com/cockroachdb/cockroach/sql/parser"
 	"github.com/cockroachdb/cockroach/sql/sqlwire"
-	"github.com/cockroachdb/cockroach/structured"
-	"github.com/cockroachdb/cockroach/util/encoding"
-	"github.com/cockroachdb/cockroach/util/log"
 )
 
 // TODO(pmattis):
@@ -197,489 +189,4 @@ func (c *conn) Send(call sqlwire.Call) (*rows, error) {
 		r.rows[i] = t
 	}
 	return r, nil
-}
-
-func (c *conn) CreateDatabase(p *parser.CreateDatabase, args []driver.Value) (*rows, error) {
-	if p.Name == "" {
-		return nil, fmt.Errorf("empty database name")
-	}
-
-	nameKey := keys.MakeNameMetadataKey(structured.RootNamespaceID, strings.ToLower(p.Name))
-	if gr, err := c.db.Get(nameKey); err != nil {
-		return nil, err
-	} else if gr.Exists() {
-		if p.IfNotExists {
-			return &rows{}, nil
-		}
-		return nil, fmt.Errorf("database \"%s\" already exists", p.Name)
-	}
-	ir, err := c.db.Inc(keys.DescIDGenerator, 1)
-	if err != nil {
-		return nil, err
-	}
-	nsID := uint32(ir.ValueInt() - 1)
-	if err := c.db.CPut(nameKey, nsID, nil); err != nil {
-		// TODO(pmattis): Need to handle if-not-exists here as well.
-		return nil, err
-	}
-	return &rows{}, nil
-}
-
-func (c *conn) CreateTable(p *parser.CreateTable, args []driver.Value) (*rows, error) {
-	if err := c.normalizeTableName(p.Table); err != nil {
-		return nil, err
-	}
-
-	dbID, err := c.lookupDatabase(p.Table.Qualifier)
-	if err != nil {
-		return nil, err
-	}
-
-	desc, err := makeTableDesc(p)
-	if err != nil {
-		return nil, err
-	}
-	if err := desc.AllocateIDs(); err != nil {
-		return nil, err
-	}
-
-	nameKey := keys.MakeNameMetadataKey(dbID, p.Table.Name)
-
-	// This isn't strictly necessary as the conditional put below will fail if
-	// the key already exists, but it seems good to avoid the table ID allocation
-	// in most cases when the table already exists.
-	if gr, err := c.db.Get(nameKey); err != nil {
-		return nil, err
-	} else if gr.Exists() {
-		if p.IfNotExists {
-			return &rows{}, nil
-		}
-		return nil, fmt.Errorf("table \"%s\" already exists", p.Table)
-	}
-
-	ir, err := c.db.Inc(keys.DescIDGenerator, 1)
-	if err != nil {
-		return nil, err
-	}
-	desc.ID = uint32(ir.ValueInt() - 1)
-
-	// TODO(pmattis): Be cognizant of error messages when this is ported to the
-	// server. The error currently returned below is likely going to be difficult
-	// to interpret.
-	err = c.db.Txn(func(txn *client.Txn) error {
-		descKey := keys.MakeDescMetadataKey(desc.ID)
-		b := &client.Batch{}
-		b.CPut(nameKey, descKey, nil)
-		b.Put(descKey, &desc)
-		return txn.Commit(b)
-	})
-	if err != nil {
-		// TODO(pmattis): Need to handle if-not-exists here as well.
-		return nil, err
-	}
-	return &rows{}, nil
-}
-
-func (c *conn) Delete(p *parser.Delete, args []driver.Value) (*rows, error) {
-	return nil, fmt.Errorf("TODO(pmattis): unimplemented: %T %s", p, p)
-}
-
-func (c *conn) Insert(p *parser.Insert, args []driver.Value) (*rows, error) {
-	desc, err := c.getTableDesc(p.Table)
-	if err != nil {
-		return nil, err
-	}
-
-	// Determine which columns we're inserting into.
-	cols, err := c.processColumns(desc, p.Columns)
-	if err != nil {
-		return nil, err
-	}
-
-	// Construct a map from column ID to the index the value appears at within a
-	// row.
-	colMap := map[uint32]int{}
-	for i, c := range cols {
-		colMap[c.ID] = i
-	}
-
-	// Verify we have at least the columns that are part of the primary key.
-	for i, id := range desc.Indexes[0].ColumnIDs {
-		if _, ok := colMap[id]; !ok {
-			return nil, fmt.Errorf("missing \"%s\" primary key column",
-				desc.Indexes[0].ColumnNames[i])
-		}
-	}
-
-	// Transform the values into a rows object. This expands SELECT statements or
-	// generates rows from the values contained within the query.
-	r, err := c.processInsertRows(p.Rows)
-	if err != nil {
-		return nil, err
-	}
-
-	b := &client.Batch{}
-	for _, row := range r.rows {
-		if len(row) != len(cols) {
-			return nil, fmt.Errorf("invalid values for columns: %d != %d", len(row), len(cols))
-		}
-		indexKey := encodeIndexKeyPrefix(desc.ID, desc.Indexes[0].ID)
-		primaryKey, err := encodeIndexKey(desc.Indexes[0], colMap, cols, row, indexKey)
-		if err != nil {
-			return nil, err
-		}
-		for i, val := range row {
-			key := encodeColumnKey(desc, cols[i], primaryKey)
-			if log.V(2) {
-				log.Infof("Put %q -> %v", key, val)
-			}
-			// TODO(pmattis): Need to convert the value type to the column type.
-			b.Put(key, val)
-		}
-	}
-	if err := c.db.Run(b); err != nil {
-		return nil, err
-	}
-
-	return &rows{}, nil
-}
-
-func (c *conn) Select(p *parser.Select, args []driver.Value) (*rows, error) {
-	if len(p.Exprs) != 1 {
-		return nil, fmt.Errorf("TODO(pmattis): unsupported select exprs: %s", p.Exprs)
-	}
-	if _, ok := p.Exprs[0].(*parser.StarExpr); !ok {
-		return nil, fmt.Errorf("TODO(pmattis): unsupported select expr: %s", p.Exprs)
-	}
-
-	if len(p.From) != 1 {
-		return nil, fmt.Errorf("TODO(pmattis): unsupported from: %s", p.From)
-	}
-	var desc *structured.TableDescriptor
-	{
-		ate, ok := p.From[0].(*parser.AliasedTableExpr)
-		if !ok {
-			return nil, fmt.Errorf("TODO(pmattis): unsupported from: %s", p.From)
-		}
-		table, ok := ate.Expr.(*parser.TableName)
-		if !ok {
-			return nil, fmt.Errorf("TODO(pmattis): unsupported from: %s", p.From)
-		}
-		var err error
-		desc, err = c.getTableDesc(table)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	// Retrieve all of the keys that start with our index key prefix.
-	startKey := proto.Key(encodeIndexKeyPrefix(desc.ID, desc.Indexes[0].ID))
-	endKey := startKey.PrefixEnd()
-	sr, err := c.db.Scan(startKey, endKey, 0)
-	if err != nil {
-		return nil, err
-	}
-
-	// All of the columns for a particular row will be grouped together. We loop
-	// over the returned key/value pairs and decode the key to extract the
-	// columns encoded within the key and the column ID. We use the column ID to
-	// lookup the column and decode the value. All of these values go into a map
-	// keyed by column name. When the index key changes we output a row
-	// containing the current values.
-	//
-	// The TODOs here are too numerous to list. This is only performing a full
-	// table scan using the primary key.
-
-	r := &rows{}
-	var primaryKey []byte
-	vals := map[string]driver.Value{}
-	for _, kv := range sr {
-		if primaryKey != nil && !bytes.HasPrefix(kv.Key, primaryKey) {
-			outputRow(r, desc.Columns, vals)
-			vals = map[string]driver.Value{}
-		}
-
-		remaining, err := decodeIndexKey(desc, desc.Indexes[0], vals, kv.Key)
-		if err != nil {
-			return nil, err
-		}
-		primaryKey = []byte(kv.Key[:len(kv.Key)-len(remaining)])
-
-		_, colID := encoding.DecodeUvarint(remaining)
-		if err != nil {
-			return nil, err
-		}
-		col, err := desc.FindColumnByID(uint32(colID))
-		if err != nil {
-			return nil, err
-		}
-		vals[col.Name] = unmarshalValue(*col, kv)
-
-		if log.V(2) {
-			log.Infof("Scan %q -> %v", kv.Key, vals[col.Name])
-		}
-	}
-
-	outputRow(r, desc.Columns, vals)
-
-	r.columns = make([]string, len(desc.Columns))
-	for i, col := range desc.Columns {
-		r.columns[i] = col.Name
-	}
-	return r, nil
-}
-
-func (c *conn) Update(p *parser.Update, args []driver.Value) (*rows, error) {
-	return nil, fmt.Errorf("TODO(pmattis): unimplemented: %T %s", p, p)
-}
-
-func (c *conn) getTableDesc(table *parser.TableName) (*structured.TableDescriptor, error) {
-	if err := c.normalizeTableName(table); err != nil {
-		return nil, err
-	}
-	dbID, err := c.lookupDatabase(table.Qualifier)
-	if err != nil {
-		return nil, err
-	}
-	gr, err := c.db.Get(keys.MakeNameMetadataKey(dbID, table.Name))
-	if err != nil {
-		return nil, err
-	}
-	if !gr.Exists() {
-		return nil, fmt.Errorf("table \"%s\" does not exist", table)
-	}
-	descKey := gr.ValueBytes()
-	desc := structured.TableDescriptor{}
-	if err := c.db.GetProto(descKey, &desc); err != nil {
-		return nil, err
-	}
-	if err := desc.Validate(); err != nil {
-		return nil, err
-	}
-	return &desc, nil
-}
-
-func (c *conn) normalizeTableName(table *parser.TableName) error {
-	if table.Qualifier == "" {
-		if c.database == "" {
-			return fmt.Errorf("no database specified")
-		}
-		table.Qualifier = c.database
-	}
-	if table.Name == "" {
-		return fmt.Errorf("empty table name: %s", table)
-	}
-	return nil
-}
-
-func (c *conn) lookupDatabase(name string) (uint32, error) {
-	nameKey := keys.MakeNameMetadataKey(structured.RootNamespaceID, name)
-	gr, err := c.db.Get(nameKey)
-	if err != nil {
-		return 0, err
-	} else if !gr.Exists() {
-		return 0, fmt.Errorf("database \"%s\" does not exist", name)
-	}
-	return uint32(gr.ValueInt()), nil
-}
-
-func (c *conn) processColumns(desc *structured.TableDescriptor,
-	node parser.Columns) ([]structured.ColumnDescriptor, error) {
-	if node == nil {
-		return desc.Columns, nil
-	}
-
-	cols := make([]structured.ColumnDescriptor, len(node))
-	for i, n := range node {
-		switch nt := n.(type) {
-		case *parser.StarExpr:
-			return c.processColumns(desc, nil)
-		case *parser.NonStarExpr:
-			switch et := nt.Expr.(type) {
-			case *parser.ColName:
-				// TODO(pmattis): If et.Qualifier is not empty, verify it matches the
-				// table name.
-				col, err := desc.FindColumnByName(et.Name)
-				if err != nil {
-					return nil, err
-				}
-				cols[i] = *col
-			default:
-				return nil, fmt.Errorf("unexpected node: %T", nt.Expr)
-			}
-		}
-	}
-
-	return cols, nil
-}
-
-func (c *conn) processInsertRows(node parser.InsertRows) (*rows, error) {
-	switch nt := node.(type) {
-	case parser.Values:
-		r := &rows{}
-		for _, row := range nt {
-			switch rt := row.(type) {
-			case parser.ValTuple:
-				var vals []driver.Value
-				for _, val := range rt {
-					switch vt := val.(type) {
-					case parser.StrVal:
-						vals = append(vals, string(vt))
-					case parser.NumVal:
-						vals = append(vals, string(vt))
-					case parser.ValArg:
-						return nil, fmt.Errorf("TODO(pmattis): unsupported node: %T", val)
-					case parser.BytesVal:
-						vals = append(vals, string(vt))
-					default:
-						return nil, fmt.Errorf("TODO(pmattis): unsupported node: %T", val)
-					}
-				}
-				r.rows = append(r.rows, vals)
-			case *parser.Subquery:
-				return nil, fmt.Errorf("TODO(pmattis): unsupported node: %T", row)
-			}
-		}
-		return r, nil
-	case *parser.Select:
-		return c.query(nt, nil)
-	case *parser.Union:
-		return c.query(nt, nil)
-	}
-	return nil, fmt.Errorf("TODO(pmattis): unsupported node: %T", node)
-}
-
-// TODO(pmattis): The key encoding and decoding routines belong in either
-// "keys" or "structured". Move them there when this code is moved to the
-// server and no longer depends on driver.Value.
-
-func encodeIndexKeyPrefix(tableID, indexID uint32) []byte {
-	var key []byte
-	key = append(key, keys.TableDataPrefix...)
-	key = encoding.EncodeUvarint(key, uint64(tableID))
-	key = encoding.EncodeUvarint(key, uint64(indexID))
-	return key
-}
-
-func encodeIndexKey(index structured.IndexDescriptor,
-	colMap map[uint32]int, cols []structured.ColumnDescriptor,
-	row []driver.Value, indexKey []byte) ([]byte, error) {
-	var key []byte
-	key = append(key, indexKey...)
-
-	for i, id := range index.ColumnIDs {
-		j, ok := colMap[id]
-		if !ok {
-			return nil, fmt.Errorf("missing \"%s\" primary key column",
-				index.ColumnNames[i])
-		}
-		// TOOD(pmattis): Need to convert the row[i] value to the type expected by
-		// the column.
-		var err error
-		key, err = encodeTableKey(key, row[j])
-		if err != nil {
-			return nil, err
-		}
-	}
-	return key, nil
-}
-
-func encodeColumnKey(desc *structured.TableDescriptor,
-	col structured.ColumnDescriptor, primaryKey []byte) []byte {
-	var key []byte
-	key = append(key, primaryKey...)
-	return encoding.EncodeUvarint(key, uint64(col.ID))
-}
-
-func encodeTableKey(b []byte, v driver.Value) ([]byte, error) {
-	switch t := v.(type) {
-	case int64:
-		return encoding.EncodeVarint(b, t), nil
-	case float64:
-		return encoding.EncodeNumericFloat(b, t), nil
-	case bool:
-		if t {
-			return encoding.EncodeVarint(b, 1), nil
-		}
-		return encoding.EncodeVarint(b, 0), nil
-	case []byte:
-		return encoding.EncodeBytes(b, t), nil
-	case string:
-		return encoding.EncodeBytes(b, []byte(t)), nil
-	case time.Time:
-		return nil, fmt.Errorf("TODO(pmattis): encode index key: time.Time")
-	}
-	return nil, fmt.Errorf("unable to encode table key: %T", v)
-}
-
-func decodeIndexKey(desc *structured.TableDescriptor,
-	index structured.IndexDescriptor, vals map[string]driver.Value, key []byte) ([]byte, error) {
-	if !bytes.HasPrefix(key, keys.TableDataPrefix) {
-		return nil, fmt.Errorf("%s: invalid key prefix: %q", desc.Name, key)
-	}
-	key = bytes.TrimPrefix(key, keys.TableDataPrefix)
-
-	var tableID uint64
-	key, tableID = encoding.DecodeUvarint(key)
-	if uint32(tableID) != desc.ID {
-		return nil, fmt.Errorf("%s: unexpected table ID: %d != %d", desc.Name, desc.ID, tableID)
-	}
-
-	var indexID uint64
-	key, indexID = encoding.DecodeUvarint(key)
-	if uint32(indexID) != index.ID {
-		return nil, fmt.Errorf("%s: unexpected index ID: %d != %d", desc.Name, index.ID, indexID)
-	}
-
-	for _, id := range index.ColumnIDs {
-		col, err := desc.FindColumnByID(id)
-		if err != nil {
-			return nil, err
-		}
-		switch col.Type.Kind {
-		case structured.ColumnType_BIT, structured.ColumnType_INT:
-			var i int64
-			key, i = encoding.DecodeVarint(key)
-			vals[col.Name] = i
-		case structured.ColumnType_FLOAT:
-			var f float64
-			key, f = encoding.DecodeNumericFloat(key)
-			vals[col.Name] = f
-		case structured.ColumnType_CHAR, structured.ColumnType_BINARY,
-			structured.ColumnType_TEXT, structured.ColumnType_BLOB:
-			var r []byte
-			key, r = encoding.DecodeBytes(key, nil)
-			vals[col.Name] = r
-		default:
-			return nil, fmt.Errorf("TODO(pmattis): decoded index key: %s", col.Type.Kind)
-		}
-	}
-
-	return key, nil
-}
-
-func outputRow(r *rows, cols []structured.ColumnDescriptor, vals map[string]driver.Value) {
-	row := make(row, len(cols))
-	for i, col := range cols {
-		row[i] = vals[col.Name]
-	}
-	r.rows = append(r.rows, row)
-}
-
-func unmarshalValue(col structured.ColumnDescriptor, kv client.KeyValue) driver.Value {
-	if !kv.Exists() {
-		return nil
-	}
-	switch col.Type.Kind {
-	case structured.ColumnType_BIT, structured.ColumnType_INT:
-		return kv.ValueInt()
-	case structured.ColumnType_FLOAT:
-		return math.Float64frombits(uint64(kv.ValueInt()))
-	case structured.ColumnType_CHAR, structured.ColumnType_BINARY,
-		structured.ColumnType_TEXT, structured.ColumnType_BLOB:
-		// TODO(pmattis): The conversion to string isn't strictly necessary, but
-		// makes log messages more readable right now.
-		return string(kv.ValueBytes())
-	}
-	return kv.ValueBytes()
 }
